+++ R/MCMC.R
@@ -190,12 +190,11 @@ Baconvergence <- function(core="MSB2K", runs=5, suggest=FALSE, verbose=TRUE, ...
 # calculate the proportion of dates that are within the age-depth model's confidence ranges
 overlap <- function(set=get('info'), digits=0, verbose=TRUE) {
   d <- set$dets[,4]
-  if(length(set$d.min) > 0)
-    d <- d[which(d >= set$d.min)]
-  if(length(set$d.max) > 0)
-    d <- d[which(d <= set$d.max)]
-  inside <- rep(1, length(d))
-  for(i in 1:length(d)) {
+  top <- ifelse(length(set$d.min) == 0, 1, min(which(d >= set$d.min)))
+  bottom <- ifelse(length(set$d.max) == 0, length(d), max(which(d <= set$d.max)))
+  these <- top:bottom
+  inside <- rep(1, length(these))
+  for(i in these) {
     daterng <- set$calib$probs[[i]]
     daterng <- cbind(cumsum(daterng[,2])/sum(daterng[,2]), daterng[,1])
     daterng <- approx(daterng[,1], daterng[,2], c((1-set$prob)/2, 1-(1-set$prob)/2))$y
@@ -205,7 +204,7 @@ overlap <- function(set=get('info'), digits=0, verbose=TRUE) {
       if(max(daterng) < min(age) || max(age) < min(daterng))
         inside[i] <- 0
   }
-  inside <- 100*sum(inside)/length(d)
+  inside <- 100*sum(inside)/length(these)
   if(verbose) 
     message(if(inside < 80) "Warning! Only ", round(inside, digits), "% of the dates overlap with the age-depth model (", 100*set$prob, "% ranges)")
 }