+++ R/calibrate.R
@@ -35,7 +35,7 @@
 #' @param up Directions of distributions if they are plotted non-mirrored. Default \code{up=TRUE}.
 #' @param BCAD The calendar scale of graphs is in \code{cal BP} by default, but can be changed to BC/AD using \code{BCAD=TRUE}.
 #' @param pch The shape of any marker to be added to the date. Defaults to a cross, \code{pch=4}. To leave empty, use \code{pch=NA}.
-#' @param cc.dir Directory where the calibration curves for C14 dates \code{cc} are located. By default \code{ccdir=c()}.
+#' @param cc.dir Directory where the calibration curves for C14 dates \code{cc} are located. By default \code{cc.dir=c()}.
 #' @author Maarten Blaauw, J. Andres Christen
 #' @return A date's distribution, added to an age-depth plot.
 #' @examples
@@ -218,20 +218,20 @@ calib.plot <- function(set=get('info'), dets=set$dets, accordion=c(), BCAD=set$B
 }
 
 
-# bacon.calib is used by the Bacon function: info$calib <- bacon.calib(dets, info, date.res, ccdir=ccdir, cutoff=cutoff)
+# bacon.calib is used by the Bacon function: info$calib <- bacon.calib(dets, info, date.res, cc.dir=cc.dir, cutoff=cutoff)
 # it calibrates C14 dates and calculate distributions for any calendar dates
 # it then returns d, cc, and probs
-bacon.calib <- function(dat, set=get('info'), date.res=100, cutoff=0.01, postbomb=set$postbomb, normal=set$normal, t.a=set$t.a, t.b=set$t.b, delta.R=set$deolta.R, delta.STD=set$delta.STD, ccdir="") {
+bacon.calib <- function(dat, set=get('info'), date.res=100, cutoff=0.01, postbomb=set$postbomb, normal=set$normal, t.a=set$t.a, t.b=set$t.b, delta.R=set$delta.R, delta.STD=set$delta.STD, cc.dir=c()) {
   # read in the curves
 
-  cc1 <- ccurve(set$cc1)
-  cc2 <- ccurve(set$cc2)
-  cc3 <- ccurve(set$cc3)
+  cc1 <- ccurve(set$cc1, cc.dir=cc.dir)
+  cc2 <- ccurve(set$cc2, cc.dir=cc.dir)
+  cc3 <- ccurve(set$cc3, cc.dir=cc.dir)
   if(set$cc4=="ConstCal" || set$cc4=="\"ConstCal\"") cc4 <- NA else
-     cc4 <- fastread(paste0(ccdir, set$cc4))[,1:3]
+     cc4 <- fastread(file.path(cc.dir, set$cc4))[,1:3] # file.path was paste0
 
   if(postbomb != 0) {
-    bomb <- ccurve(postbomb, postbomb=TRUE, glue=FALSE) # glue=FALSE added July 2023
+    bomb <- ccurve(postbomb, postbomb=TRUE, glue=FALSE, cc.dir=cc.dir) # glue=FALSE added July 2023
     # bomb.x <- seq(max(bomb[,1]), min(bomb[,1]), by=-.1) # interpolate
     bomb <- bomb[order(bomb[,1], decreasing=FALSE),]
     bomb.x <- seq(min(bomb[,1]), max(bomb[,1]), by=.1) # interpolate